Uplimit Course Submission
The script generates a personalized newsletter every week that summarizes each podcast episode released in that week. It includes information about the guest on the episode, the main topics discussed as well as some highlights. It works by collecting a list of RSS feeds from the user and on a periodic basis, processes the latest episodes and generates the newsletter. This acts as a round-up of the week, provides the user with the right level of detail that allows them to decide which of the episodes are appealing and must be listened to.